+++ Tests/Westwind.Utilities.Test/AppConfiguration/CustomConfigFileConfigurationTests.cs
@@ -47,6 +47,8 @@ namespace Westwind.Utilities.Configuration.Tests
         [TestMethod]
         public void WriteConfigurationTest()
         {
+            File.Delete(TestHelpers.GetTestConfigFilePath());
+
             var config = new CustomConfigFileConfiguration();
             config.Initialize();
             
@@ -60,10 +62,22 @@ namespace Westwind.Utilities.Configuration.Tests
             config.AppConnectionString = "server=.;database=unsecured";
 
             // Complex Types
-            config.ComplexType.Company = "Updated Company";
+            config.License.Company = "Updated Company";
             config.ServerList[0] = "UpdatedServerName";
 
+            config.License.Name = "Rick";
+            config.License.Company = "West Wind 2";
+            config.License.LicenseKey = "RickWestWind2-51231223";
+
             config.Write();
+
+            config = null;
+            config = new CustomConfigFileConfiguration();
+            config.Initialize();
+
+            Console.WriteLine(config.License.LicenseKey);
+            Assert.IsTrue(config.License.LicenseKey == "RickWestWind2-51231223");
+            
             
             string text = File.ReadAllText(TestHelpers.GetTestConfigFilePath());
             Console.WriteLine(text);
@@ -77,7 +91,7 @@ namespace Westwind.Utilities.Configuration.Tests
             Assert.IsTrue(text.Contains(@"<add key=""AppConnectionString"" value=""z6+T5mzXbtJBEgWqpQNYbBss0csbtw2b/qdge7PUixE="" />"));
 
             // Complex Value
-            Assert.IsTrue(text.Contains(@"Updated Company"));
+            Assert.IsTrue(text.Contains(@"West Wind 2"));
 
             // List values
             Assert.IsTrue(text.Contains(@"<add key=""ServerList1"""));
@@ -102,7 +116,7 @@ namespace Westwind.Utilities.Configuration.Tests
             config.Password = "seekrit2";
             config.AppConnectionString = "server=.;database=unsecured";
 
-            config.ComplexType.Company = "Updated Company";
+            config.License.Company = "Updated Company";
             config.ServerList[0] = "UpdatedServerName";
 
             config.Write();
@@ -130,7 +144,7 @@ namespace Westwind.Utilities.Configuration.Tests
             config2.Initialize();
             config2.Read();
 
-            Assert.IsTrue(config2.ComplexType.Company == "Updated Company");
+            Assert.IsTrue(config2.License.Company == "Updated Company");
             Assert.IsTrue(config2.ServerList[0] == "UpdatedServerName");
 
         }