+++ Tests/Westwind.Utilities.Test/AppConfiguration/JsonFileConfigurationTests.cs
@@ -63,6 +63,8 @@ namespace Westwind.Utilities.Configuration.Tests
         [TestMethod]
         public void WriteEncryptedConfigurationTest()
         {
+            File.Delete(STR_JSONCONFIGFILE);
+
             var config = new JsonFileConfiguration();
             config.Initialize(STR_JSONCONFIGFILE);
 
@@ -70,7 +72,18 @@ namespace Westwind.Utilities.Configuration.Tests
             config.Password = "seekrit2";
             config.AppConnectionString = "server=.;database=unsecured";
 
+            config.License.Company = "West Wind 2";
+            config.License.LicenseKey = "RickWestWind2-51123222";
+
             config.Write();
+
+            // completely reload settings
+            config = new JsonFileConfiguration();
+            config.Initialize(STR_JSONCONFIGFILE);
+
+            Assert.IsTrue(config.Password == "seekrit2");
+            Assert.IsTrue(config.License.Company == "West Wind 2");
+            Assert.IsTrue(config.License.LicenseKey == "RickWestWind2-51123222");
 
             string jsonConfig = File.ReadAllText(STR_JSONCONFIGFILE);
             Console.WriteLine(jsonConfig);